requirements.txt
streamlit
python-dotenv
transformers
huggingface-hub
google-generativeai
matplotlib
pandas
langdetect
geopy
googletrans==4.0.0-rc1
sqlite3
